software: kernel: Support compilation of project-specific modules
Involved kernel modules are listed under a variable named
`modules-$(ARCH)-phy_os-y` or `modules-$(ARCH)-virt-y`,
in which $(ARCH) is explicitly substituted with the ISA name.

For example, in SERVE project, a variable named
`modules-riscv-phy_os-y` is defined.
When Keystone Enclave is enabled in one SERVE platform,
the keystone-driver module would be compiled.